requirements.txt
Flask==2.3.3
pandas==2.1.1
scikit-learn==1.3.1
joblib==1.3.2
numpy==1.24.3
xgboost==1.7.6
imbalanced-learn==0.11.0

# Security and Rate Limiting
Flask-Limiter==3.5.0
Werkzeug==2.3.7

# Database Dependencies
mysql-connector-python==8.2.0
SQLAlchemy==2.0.23
pymongo==4.6.0

# Environment and Configuration
python-dotenv==1.0.0

# Compliance and Scheduling
schedule==1.2.0
APScheduler==3.10.4

# Monitoring and Logging
prometheus-flask-exporter==0.23.0

# Data Processing
scipy==1.11.3

# Enhanced ML Training Dependencies
lightgbm==4.1.0
catboost==1.2
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Advanced ML and Data Science
optuna==3.4.0
shap==0.43.0
feature-engine==1.6.2
category-encoders==2.6.1

# Model Evaluation and Validation
yellowbrick==1.5
evidently==0.4.9

# Security and JWT
PyJWT==2.8.0
cryptography==41.0.7
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1

# Optional for enhanced security
# flask-wtf==1.1.1
# flask-talisman==1.1.0

# Development dependencies (optional)
# pytest==7.4.2
# pytest-flask==1.2.0
# black==23.9.1
# flake8==6.1.0
# jupyter==1.0.0
# ipykernel==6.25.2
